Update property-effects.html
When an computed property or complex observer is triggered and the calling element has a property 'usePolymer1StyleObserver' set to true , require all arguments to be defined to call the function. Elements can set their v1 hybrid elements to have this property in order to more incrementally upgrade to Polymer2. This is meant to be in Polymer2 only and not Polymer3, as a aid to incrementally update apps that will have many changes to their observers.